studentList

+++ model/model.go
@@ -1,15 +1,15 @@
 package model
 
 type Student struct {
-	Id uint8
-	No int8
+	Id uint32
+	No uint32
 	Name string
 	CScore float32
 	MathScore float32
 	EnglishScore float32
 	TotalScore float32
 	AverageScore float32
-	Ranking int8
+	Ranking uint32
 	CreatedTime string
 	UpdatedTime string
 }
